Adjust WinRT property hierarchy
